moves food methods to controller
